pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-memory"
version = "0.1.0"
description = "CosmosDB-backed agent memory with multi-tier storage"
requires-python = ">=3.12"
dependencies = [
    "agent-framework==1.0.0b251028",
    "openai>=2.5.0",
    "python-dotenv>=1.1.1",
    "azure-cosmos==4.14.0",
    "azure-identity>=1.19.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]

[tool.setuptools.packages.find]
include = ["memory*", "agent*", "application*", "demo*", "cosmos_db_search*"]

[tool.uv]
prerelease = "allow"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
